requirements.txt
# Core API
fastapi[standard]==0.135.1
uvicorn[standard]==0.41.0

# Database
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4

# Pydantic (modern v2)
pydantic==2.12.5
pydantic-settings==2.3.4

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.22

# File parsing
PyPDF2==3.0.1
pandas==3.0.1
openpyxl==3.1.5

# LLM
groq==0.30.0

# Testing
httpx==0.28.1
pytest==8.3.2
